fix: Fix frame drop(s) in main-thread by switching the work of audio switching off the main-thread
